+++ elasticsearch7/search-index/src/main/java/org/vertexium/elasticsearch7/Elasticsearch7SearchIndex.java
@@ -626,6 +626,15 @@ public class Elasticsearch7SearchIndex implements SearchIndex, SearchIndexWithVe
 
             Map<String, Object> fieldsToSet = getPropertiesAsFields(graph, element.getProperties());
 
+            for (Property property : element.getProperties()) {
+                for (Visibility hiddenVisibility : property.getHiddenVisibilities()) {
+                    String hiddenVisibilityPropertyName = addVisibilityToPropertyName(graph, HIDDEN_PROPERTY_FIELD_NAME, hiddenVisibility);
+                    if (!isPropertyInIndex(graph, HIDDEN_PROPERTY_FIELD_NAME, hiddenVisibility)) {
+                        addPropertyToIndex(graph, indexInfo, hiddenVisibilityPropertyName, hiddenVisibility, Boolean.class, false, false, false);
+                    }
+                    fieldsToSet.put(hiddenVisibilityPropertyName, true);
+                }
+            }
             for (Visibility hiddenVisibility : element.getHiddenVisibilities()) {
                 String hiddenVisibilityPropertyName = addVisibilityToPropertyName(graph, HIDDEN_VERTEX_FIELD_NAME, hiddenVisibility);
                 if (!isPropertyInIndex(graph, HIDDEN_VERTEX_FIELD_NAME, hiddenVisibility)) {
@@ -1591,6 +1600,11 @@ public class Elasticsearch7SearchIndex implements SearchIndex, SearchIndexWithVe
 
     @Override
     public boolean isFieldLevelSecuritySupported() {
+        return true;
+    }
+
+    @Override
+    public boolean isDeleteElementSupported() {
         return true;
     }